Message update: auto-add missing keys
If a key (that is in default.yml) is missing from a message catalog,
add that key/value to the catalog, and also add the key to a
"NEEDS_TRANSLATE" list in the catalog.  This makes it easy for
translators to see what new keys they need to add if default.yml is
updated.