AI thread safety change.
ChessAI no longer directly redraws the board (via Game.doMove())
because that leads to Bukkit/Minecraft API calls which are not
thread-safe.  Instead it just sets a flag in the Game instances which
indicates the AI move has been done.  This is picked up by the ticker
task in the main thread.